+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Package class list lookup now using ClassHelper.getDefaultClassLoader().loadClass().
+
 ## [1.3.3] - 2024-04-01
 
 ### Added